4.ObstacleAvoidance/requirements.txt
flask>=2.0.1
flask-socketio>=5.1.1
eventlet>=0.33.0
RPi.GPIO>=0.7.1 